+++ config/me_cms.php
@@ -7,8 +7,6 @@ return ['MeCms' => [
      ],
     //Default layout
     'default' => [
-        //Displays the alert for the cookie policy
-        'cookies_policy' => true,
         //"Contact us" form (enabled or disabled)
         'contact_us' => true,
         //Facebook app ID or `false`
@@ -21,6 +19,10 @@ return ['MeCms' => [
         'offline' => false,
         //Text to display when the site is offline or `false`
         'offline_text' => false,
+        //Other application css files to load automatically (default `contents` and `layout`)
+        'other_css' => ['contents', 'layout'],
+        //Other application script files to load automatically (default empty)
+        'other_js' => [],
         //Number of records to show per page
         'records' => 10,
         //Number of records to show on RSS